Added default values for Icon, moved TransactionForm to components directory
